requirements.txt
prompt_toolkit
requests
tabulate
boto3
botocore
websocket_client

